applications/image-processing/aws/requirements.txt
numpy==1.16.3
pillow==3.1.2
